add correct movemant

+++ src/main/com/testenv/models/Block.java
@@ -5,7 +5,7 @@ import javafx.scene.paint.Paint;
 
 public class Block extends RectangleUnit {
 
-    public Block(int x, int y, Size size, double angle, Paint color) {
+    public Block(double x, double y, Size size, double angle, Paint color) {
         super(x, y, size, angle, color);
     }
 }
